+++ R/shapes.R
@@ -18,6 +18,10 @@
 #'   Default = "sf", which uses \code{sf::st_read}. Currenlty, you can also
 #'   specify "rgdal", which reads the file using rgdal::readOGR.
 #'   To just return the file paths for the files use method = "zip".
+#' @param quiet_parse Whether to download file quietly. Passed to
+#'   [`download_file()`]. Default is `FALSE`.
+#' @param quiet_parse Whether to read boundaries dataset quietly. Applies to
+#'   `method = "sf"`. Default is `TRUE`.
 #'
 #' @details Downloads the spatial boundaries from the DHS spatial repository,
 #'   which can be found at \url{https://spatialdata.dhsprogram.com/home/}.
@@ -42,7 +46,9 @@
 download_boundaries <- function(surveyNum=NULL,
                                 surveyId = NULL,
                                 countryId = NULL,
-                                method = "sf"){
+                                method = "sf",
+                                quiet_download = FALSE,
+                                quiet_parse = TRUE){
 
   # helper funcs
   build_final_url <- function(jobId) {
@@ -109,7 +115,7 @@ download_boundaries <- function(surveyNum=NULL,
 
   # download the shape file and read it in
   tf2 <- tempfile()
-  file <- download.file(url, tf2)
+  file <- download.file(url, tf2, quiet = quiet_download)
   unzipped_files <- suppressWarnings(unzip(tf2, exdir = tempfile()))
   file <- grep("dbf", unzipped_files, value=TRUE)
 
@@ -127,7 +133,7 @@ download_boundaries <- function(surveyNum=NULL,
 
     # here if we want to add more read in options
     if(method == "sf") {
-      res <- lapply(file, sf::st_read)
+      res <- lapply(file, sf::st_read, quiet = quiet_parse)
       names(res) <- vapply(file,
                            function(x) { sf::st_layers(x)$name },
                            character(1))